+++ src/shared/contract/customer-contract/command_release_credit.go
@@ -1,6 +1,6 @@
 package customercontract
 
 type ReleaseCreditCommand struct {
-	CustomerID   int `json:"customer_id"`
-	CreditAmount int `json:"credit_amount"`
+	CustomerID   int64 `json:"customer_id"`
+	CreditAmount int   `json:"credit_amount"`
 }
